+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.20.1](///compare/v2.20.0...v2.20.1) (2026-01-05)
+
+## [2.20.0](///compare/v2.19.0...v2.20.0) (2026-01-05)
+
+
+### Features
+
+* **extension:** add YouTube quiz generation to description c241476
+* **ui:** implement grid layout for video extraction 4def7d8
+* **video:** add YouTube quiz generator button bf83a95
+* **youtube:** add YouTube quiz generator module a4ef54a
+
 ## [2.19.0](///compare/v2.18.0...v2.19.0) (2026-01-05)